+++ components/Checkbox.tsx
@@ -1,3 +1,5 @@
+'use client'
+
 import { type InputHTMLAttributes, forwardRef, useId } from 'react'
 import { cn } from '@/lib/utils'